v0.0.2 Integer to Decimal for business

+++ database/migrations/0001_01_01_000003_create_akun_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->unsignedBigInteger('kode')->unique();
             $table->string('nama');
             $table->enum('kategori', ['aset','beban','utang','modal','pendapatan']);
-            $table->integer('debit')->default(0);
-            $table->integer('kredit')->default(0);
+            $table->decimal('debit', 15, 2)->default(0);
+            $table->decimal('kredit', 15, 2)->default(0);
             $table->boolean('aktif')->default(1);
             $table->softDeletes('deleted_at', precision:0);
             $table->timestamp('created_at')->useCurrent();
